+++ CHANGELOG.md
@@ -3,4 +3,4 @@ Changelog
 
 ### 3.2.2.7
 
--
+- Fix issues with filesizes not showing correctly when downloading mods for a pack